# Lower IO buffer sizes and max compressed block size (better cache locality)