--- a/tests/queries/0_stateless/01598_memory_limit_zeros.sql
+++ b/tests/queries/0_stateless/01598_memory_limit_zeros.sql
@@ -1,2 +1,2 @@
-SET max_memory_usage = 1, max_untracked_memory = 1000000;
+SET max_memory_usage = 1, max_untracked_memory = 500000;
 select 'test', count(*) from zeros_mt(1000000) where not ignore(zero); -- { serverError 241 }