Changelog category (leave one):
- Performance Improvement


Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Lower IO buffer sizes and max compressed block size for better cache locality. Note that some datasets may compress worse with these settings.