@alexey-milovidov 2021-07-29
Lowered memory usage as a bonus.