@alexey-milovidov 2021-07-29
@azat Integration test `test_distributed_directory_monitor_split_batch_on_failure` maybe depends on subtle differences in memory usage.