@alexey-milovidov 2021-07-29
Up to 5 times performance increase and 2 times decrease on specific queries.