@azat 2021-07-30
@alexey-milovidov 

New limits:
- `01730_distributed_group_by_no_merge_order_by_long` - `82Mi` (checked only w/o sanitizers)
- `test_distributed_directory_monitor_split_batch_on_failure` - `25Mi`