@alexey-milovidov 2021-07-30
Ok. Now what is remaining:

1. Test the most regressed queries in performance test.
Although performance test is already good in overall.

2. Test on a server with HDD.
With queries just after dropping cache and queries with direct IO.

3. Test on a VM with EBS volume (network disk in Yandex.Cloud).
With queries just after dropping cache and queries with direct IO.