@azat 2021-07-30
@alexey-milovidov some changes for other tests:
- `01641_memory_tracking_insert_optimize`: `max_memory_usage='6Mi'`
- `01598_memory_limit_zeros`: `max_untracked_memory='500Ki'` (note, that failed allocation is still from the ZerosSource::createColumn, not in some basic buffers for query, or similar)