@alexey-milovidov 2021-07-31
The most regressed query is actually faster when I test on machine:

```
clickhouse-benchmark <<< "SELECT RegionID, uniq(UserID) AS u FROM test.hits GROUP BY RegionID ORDER BY u DESC LIMIT 10"

New:

  20,62%  [kernel]                               [k] acpi_processor_ffh_cstate_enter
  16,73%  clickhouse                             [.] UniquesHashSet<DefaultHash<unsigned long, void> >::insert
   8,12%  clickhouse                             [.] DB::Aggregator::executeImplBatch<false, false, DB::AggregationMethodOneNumber<unsigned int, HashMapTable<unsigned long, HashMapCell<unsigned long, char*, HashCRC32<unsigned long>, HashTableNoSt
   4,48%  [kernel]                               [k] acpi_idle_do_entry
   3,67%  clickhouse                             [.] DB::IAggregateFunctionHelper<DB::AggregateFunctionUniq<unsigned long, DB::AggregateFunctionUniqUniquesHashSetData> >::addBatch
   2,53%  clickhouse                             [.] LZ4::(anonymous namespace)::decompressImpl<32ul, false>
   2,20%  clickhouse                             [.] LZ4::(anonymous namespace)::decompressImpl<16ul, true>
   2,06%  clickhouse                             [.] LZ4::(anonymous namespace)::decompressImpl<8ul, true>
   2,05%  clickhouse                             [.] LZ4::(anonymous namespace)::decompressImpl<16ul, false>
   1,97%  clickhouse                             [.] UniquesHashSet<DefaultHash<unsigned long, void> >::resize
   1,72%  clickhouse                             [.] UniquesHashSet<DefaultHash<unsigned long, void> >::merge
   1,15%  [kernel]                               [k] copy_user_generic_string
   0,82%  clickhouse                             [.] memcpy

Queries executed: 4092.

localhost:9000, queries 4092, QPS: 33.366, RPS: 296088778.352, MiB/s: 3388.467, result RPS: 333.663, result MiB/s: 0.004.

0.000%          0.026 sec.
10.000%         0.029 sec.
20.000%         0.029 sec.
30.000%         0.029 sec.
40.000%         0.030 sec.
50.000%         0.030 sec.
60.000%         0.030 sec.
70.000%         0.030 sec.
80.000%         0.031 sec.
90.000%         0.032 sec.
95.000%         0.032 sec.
99.000%         0.034 sec.
99.900%         0.037 sec.
99.990%         0.038 sec.


Old:

  20,41%  [kernel]                           [k] acpi_processor_ffh_cstate_enter
  16,89%  clickhouse                         [.] UniquesHashSet<DefaultHash<unsigned long, void> >::insert
   8,28%  clickhouse                         [.] DB::Aggregator::executeImplBatch<false, false, DB::AggregationMethodOneNumber<unsigned int, HashMapTable<unsigned long, HashMapCell<unsigned long, char*, HashCRC32<unsigned long>, HashTableNoState>
   6,90%  [kernel]                           [k] copy_user_generic_string
   4,94%  [kernel]                           [k] acpi_idle_do_entry
   3,68%  clickhouse                         [.] DB::IAggregateFunctionHelper<DB::AggregateFunctionUniq<unsigned long, DB::AggregateFunctionUniqUniquesHashSetData> >::addBatch
   2,66%  clickhouse                         [.] LZ4::(anonymous namespace)::decompressImpl<32ul, false>
   2,33%  clickhouse                         [.] LZ4::(anonymous namespace)::decompressImpl<16ul, true>
   2,18%  clickhouse                         [.] LZ4::(anonymous namespace)::decompressImpl<16ul, false>
   2,15%  clickhouse                         [.] LZ4::(anonymous namespace)::decompressImpl<8ul, true>
   2,02%  clickhouse                         [.] UniquesHashSet<DefaultHash<unsigned long, void> >::resize
   1,73%  clickhouse                         [.] UniquesHashSet<DefaultHash<unsigned long, void> >::merge
   0,88%  clickhouse                         [.] memcpy
   0,66%  clickhouse                         [.] CityHash_v1_0_2::CityHash128WithSeed

Queries executed: 4237.

localhost:9000, queries 4237, QPS: 32.060, RPS: 284494471.402, MiB/s: 3255.781, result RPS: 320.597, result MiB/s: 0.004.

0.000%          0.027 sec.
10.000%         0.030 sec.
20.000%         0.030 sec.
30.000%         0.030 sec.
40.000%         0.031 sec.
50.000%         0.031 sec.
60.000%         0.031 sec.
70.000%         0.032 sec.
80.000%         0.032 sec.
90.000%         0.033 sec.
95.000%         0.034 sec.
99.000%         0.036 sec.
99.900%         0.039 sec.
99.990%         0.089 sec.
```

The function `copy_user_generic_string` (copying data from page cache to userspace) becomes much more cheap due to cache locality.