@alexey-milovidov 2021-07-31
Major performance degradation on a server with slow HDD RAID:

```
Before:

$ for i in {1..10}; do echo 3 | sudo tee /proc/sys/vm/drop_caches && clickhouse-client --time --query "SELECT sum(ignore(*)) FROM test.hits"; done3
0
12.185
3
0
12.616
3
0
12.473
3
0
12.827
3
0
12.854
3
0
11.613
3
0
11.738
3
0
12.385
3
0
13.205
3
0
13.629

After:

$ for i in {1..10}; do echo 3 | sudo tee /proc/sys/vm/drop_caches && clickhouse-client --time --query "SELECT sum(ignore(*)) FROM test.hits"; done
3
0
24.009
3
0
20.183
3
0
21.827
3
0
19.322
3
0
19.643
3
0
20.333
3
0
18.532
3
0
20.282
3
0
17.893
3
0
19.576
```

Not ready for merge.