@alexey-milovidov 2021-08-05
We need to adjust buffer size automatically.

Introduce `min_read_buffer_size` setting, set it to 64K by default.

Determine if block device under filesystem is rotational or not and its optimal IO size:

1. Found mount point of the filesystem with `statvfs`.
2. Read `/proc/self/mountinfo` and find a row with this mount point.
3. Extract device number, e.g. `9:2` from `74 21 9:2 / /opt rw,noatime shared:29 - ext4 /dev/md2 rw,nobarrier,stripe=1792,data=ordered`.
4. Read `/sys/dev/block/9:2/queue/optimal_io_size` and `/sys/dev/block/9:2/queue/rotational`.

If `optimal_io_size` is non zero, cap it with `min_read_buffer_size` and `max_read_buffer_size` - this is the buffer size to use.
If it is zero and device is rotational, choose `max_read_buffer_size`, otherwise `min_read_buffer_size`.

@azat Comments?
Maybe there is more simple way to check if we are working with HDD?