@Algunenano 2021-08-06
> Extract device number, e.g. 9:2 from 74 21 9:2 / /opt rw,noatime shared:29 - ext4 /dev/md2 rw,nobarrier,stripe=1792,data=ordered.
> Read /sys/dev/block/9:2/queue/optimal_io_size and /sys/dev/block/9:2/queue/rotational.

There is a missing step there which is to go from partition to disk/device if needed. At least in my machine (5.13.8) those values are only available per disk, nor per partition.